+++ sdk/servicebus/azure-servicebus/azure/servicebus/__init__.py
@@ -12,7 +12,6 @@ from ._servicebus_client import ServiceBusClient
 from ._servicebus_sender import ServiceBusSender
 from ._servicebus_receiver import ServiceBusReceiver
 from ._servicebus_session_receiver import ServiceBusSessionReceiver
-from ._servicebus_session import ServiceBusSession
 from ._common.message import Message, BatchMessage, PeekedMessage, ReceivedMessage
 from ._common.constants import ReceiveMode, SubQueue
 from ._common.auto_lock_renewer import AutoLockRenew
@@ -29,7 +28,6 @@ __all__ = [
     'ServiceBusClient',
     'ServiceBusReceiver',
     'ServiceBusSessionReceiver',
-    'ServiceBusSession',
     'ServiceBusSender',
     'TransportType',
     'AutoLockRenew'