+++ sdk/servicebus/azure-servicebus/azure/servicebus/aio/__init__.py
@@ -7,7 +7,6 @@ from ._async_message import ReceivedMessage
 from ._servicebus_sender_async import ServiceBusSender
 from ._servicebus_receiver_async import ServiceBusReceiver
 from ._servicebus_session_receiver_async import ServiceBusSessionReceiver
-from ._servicebus_session_async import ServiceBusSession
 from ._servicebus_client_async import ServiceBusClient
 from ._async_auto_lock_renewer import AutoLockRenew
 
@@ -17,6 +16,5 @@ __all__ = [
     'ServiceBusSender',
     'ServiceBusReceiver',
     'ServiceBusSessionReceiver',
-    'ServiceBusSession',
     'AutoLockRenew'
 ]